refactor: change folder structure and rename files
