+++ src/models/albums/albumsService.ts
@@ -1,10 +1,10 @@
 import { RESTDataSource, RequestOptions } from 'apollo-datasource-rest';
-import { API, AlbumsAPIEndpoint } from '../const/api';
-import { Album, CreateAlbumArgs, UpdateAlbumArgs } from '../types/albums';
-import { DeleteResponse } from '../types/deleteResponse';
-import { List } from '../types/list';
-import { QueryParams } from '../types/queryParams';
-import getQueryParams from '../utils/getQueryParams';
+import { API, AlbumsAPIEndpoint } from '../../const/api';
+import { Album, CreateAlbumArgs, UpdateAlbumArgs } from '../../types/albums';
+import { DeleteResponse } from '../../types/deleteResponse';
+import { List } from '../../types/list';
+import { QueryParams } from '../../types/queryParams';
+import getQueryParams from '../../utils/getQueryParams';
 
 export interface AlbumsAPIDataSource {
   getAlbum: (id: string) => Promise<Album>
