+++ src/models/artists/ArtistsTypeDefs.ts
