+++ src/models/artists/artistsService.ts
@@ -1,10 +1,10 @@
 import { RESTDataSource, RequestOptions } from 'apollo-datasource-rest';
-import { API, ArtistsAPIEndpoint } from '../const/api';
-import { Artist, CreateArtistArgs, UpdateArtistArgs } from '../types/artists';
-import { DeleteResponse } from '../types/deleteResponse';
-import { List } from '../types/list';
-import { QueryParams } from '../types/queryParams';
-import getQueryParams from '../utils/getQueryParams';
+import { API, ArtistsAPIEndpoint } from '../../const/api';
+import { Artist, CreateArtistArgs, UpdateArtistArgs } from '../../types/artists';
+import { DeleteResponse } from '../../types/deleteResponse';
+import { List } from '../../types/list';
+import { QueryParams } from '../../types/queryParams';
+import getQueryParams from '../../utils/getQueryParams';
 
 export interface ArtistsAPIDataSource {
   getArtist: (id: string) => Promise<Artist>
