+++ src/models/bands/BandsResolver.ts
