+++ src/models/favourites/favouritesService.ts
@@ -1,8 +1,8 @@
 import { RESTDataSource, RequestOptions } from 'apollo-datasource-rest';
-import { API, FavouritesAPIEndpoint } from '../const/api';
-import { Favourites } from '../types/favourites';
-import { QueryParams } from '../types/queryParams';
-import getQueryParams from '../utils/getQueryParams';
+import { API, FavouritesAPIEndpoint } from '../../const/api';
+import { Favourites } from '../../types/favourites';
+import { QueryParams } from '../../types/queryParams';
+import getQueryParams from '../../utils/getQueryParams';
 
 export interface FavouritesAPIDataSource {
   getFavourites: (queryParams?: QueryParams) => Promise<Favourites>;
